move pytest fixtures to specific module for nilearn.image functions
